requirements.txt
pandas
numpy
matplotlib
scipy
PyWavelets
Pillow
torch
scikit-learn
seaborn